refactor(store): simplify Store namespace usage across files
