+++ core/Common/Store/Stores.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Testo\Common\Store;
 
+use Testo\Common\Store;
+
 /**
  * Registry of stores — the single entry point for opening persistent, cross-run storage.
  *
